size_t -> int

+++ main.cpp
@@ -2,19 +2,19 @@
 #include <fstream>
 
 /**
- * The program reads numbers from a file and prsize_ts the sum of them
+ * The program reads numbers from a file and prints the sum of them
  * Example: main.exe ./tests/input.in
  * @param  argc [description]
  * @param  argv [description]
  * @return      [description]
  */
 int main(int argc, const char* argv[]) {
-  size_t a = 0;
-  size_t sum = 0;
-  size_t avg = 0;
-  size_t arraySize = 1;
-  size_t numbersAdded = 0;
-  size_t* numbers = new size_t[arraySize];
+  int a = 0;
+  int sum = 0;
+  int avg = 0;
+  int arraySize = 1;
+  int numbersAdded = 0;
+  int* numbers = new int[arraySize];
 
   std::ifstream myReadFile(argv[1]);
 
@@ -23,9 +23,9 @@ int main(int argc, const char* argv[]) {
   while (myReadFile >> a) {
     /* Expand array if numbers added is greater or equals size of array */ 
     if (numbersAdded >= arraySize) {
-      size_t *tmp = new size_t[arraySize + 1];
+      int *tmp = new int[arraySize + 1];
 
-      for (size_t i = 0; i < arraySize; i++) {
+      for (int i = 0; i < arraySize; i++) {
         tmp[i] = numbers[i];
       }
 
@@ -46,12 +46,12 @@ int main(int argc, const char* argv[]) {
   avg = sum / arraySize;
 
   std::cout << "Input: "; 
-  for (size_t i = 0; i < arraySize; i++) {
+  for (int i = 0; i < arraySize; i++) {
     std::cout << numbers[i] << '\t';
   }
 
   std::cout << "\tOutput: ";
-  for (size_t i = 0; i < arraySize; i++) {
+  for (int i = 0; i < arraySize; i++) {
     (numbers[i] > avg) ? std::cout << numbers[i] << '\t': std::cout << "";
   }
